Created a new BaseController #65
Created a BaseController class because this logic is reused everywhere
in the application.

This is the start of setting up the user’s profile with name and
password change.
