+++ app/default/templates/default.html
@@ -1,22 +1,23 @@
 <div class="jumbotron">
   <h1 class="header">Get Pumped and Make Gains</h1>
   <div class="row">
-    <div class="col-md-6">
+    <div class="col-md-8">
       <p>Record, Recall, Report</p>
       <p>
         There's nothing more annoying than trying to remember your last session in
         the gym when all you want to do is get a sick pump.
       </p>
+      <p>Start your session</p>
+      <p>Look up your workout</p>
+      <p>Re-run your last set</p>
       <p>
-        Getting huge is as simple as lifting heavier weights each time you hit the
-        gym.
+        <a href="login" class="btn btn-primary btn-lg login-button">
+          Get Started
+        </a>
       </p>
     </div>
-    <div class="col-md-6">
+    <div class="col-md-4">
       <img src="/img/last-workout.png" style="width: 100%">
     </div>
   </div>
-  <a href="login" class="btn btn-primary btn-lg login-button">
-    Get Started
-  </a>
 </div>
