+++ app/exercises/router.js
@@ -1,11 +1,10 @@
-import _ from 'underscore';
-import Marionette from 'backbone.marionette';
-
-import root from '../root';
+import {BaseController} from '../base/controller';
 
 import {ExerciseLayout} from './views/base';
 
-export const Controller = Marionette.Object.extend({
+export const Controller = BaseController.extend({
+  layoutView: ExerciseLayout,
+
   listExercise: function() {
     const layout = this.showAndGetLayout();
     layout.showExerciseList();
@@ -13,26 +12,7 @@ export const Controller = Marionette.Object.extend({
 
   showExercise: function() {
 
-  },
-
-  showingMyView: function() {
-    const view = root.getChildView('main');
-    const cid = this.getOption('indexCid');
-    if (_.isUndefined(cid)) {
-      return false;
-    }
-    return view.cid === cid;
-  },
-
-  showAndGetLayout: function() {
-    if (this.showingMyView()) {
-      return root.getChildView('main');
-    }
-    const layout = new ExerciseLayout();
-    root.showChildView('main', layout);
-    return layout;
   }
-
 });
 
 export const routes = {
