+++ app/templates/nav.html
@@ -12,11 +12,16 @@
     </a>
   </p>
 </div>
+
 <ul class="nav navbar-nav hidden-xs">
   <li><a href="/exercise/" class="authenticated exercise">Exercises</a></li>
   <li><a href="/workout/" class="authenticated workout">Workouts</a></li>
 </ul>
 <ul class="nav navbar-nav navbar-right hidden-xs">
+  <li><a href="/profile/" class="authenticated profile">
+    My Profile
+    <i class="glyphicon glyphicon-user"></i>
+  </a></li>
   <li><a href="/logout" class="authenticated text-muted logout">
     Logout
     <i class="glyphicon glyphicon-off"></i>
