+++ app/user/router.js
@@ -1,0 +1,10 @@
+import Marionette from 'backbone.marionette';
+
+import {BaseController} from '../base/controller';
+
+export const Controller = BaseController.extend({
+
+});
+
+export const routes = {
+};
