+++ assets/index.html
@@ -1,13 +1,14 @@
 <!DOCTYPE html>
 <head lang="en">
   <link href="//fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-
-  <link href="/css/bootstrap.min.css" rel="stylesheet" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+  <link rel="shortcut icon" href="/img/favicon.ico">
+  <link href="/css/bootstrap.min.css" rel="stylesheet">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
 
   <title>Get PUMP3D</title>
 </head>
 <body>
+  <noscript>Please enable JavaScript!</noscript>
   <div id="root"></div>
   <script src="/js/bundle.js"></script>
 </body>
